Algorithm change for Hummingbird
In the case that Hummingbird doesn't find anything close, it the bot
will now return None for Hummingbird instead of the first anime on
Hummingbird's search list. This is because Hummingbird's API results
tend to be broad and will generally return *something* instead of
nothing. Linking nothing is preferable to linking the wrong thing,
especially when the other databases may be correct.