Further search algorithm improvements
Refined the search algorithm for finding the correct show on
Hummingbird. If it fails to find something using the Anilist romaji
title, it will also perform a search using all of Anilist's synonyms, as
well as the listed English title. Failing that, it will now also check
the romaji *and* English titles from MAL (instead of just the MAL romaji
title).