Import Config unconditionally
Trying to import the config and failing doesn't protect against failures
to use the config definitions in later code, so it won't help to ignore
the ImportError. The whole try block can be removed.

Additionally, the Config module was already imported above, so the import
inside the try block is redundant.