Improve tag detection in PM processing
Before doing additional processing, checking for a tag in a PM uses
expressions like `('{' and '}') in msg.body`, but since the expression
`('{' and '}')` just evaluates to `'}'` this only checks for closing
delimiters.

Additionally, the conditional can be inverted to continue and the
followup actions can be flattened out as a result.